Execute as per thread configuration in test files (#27)
* Execute as per thread configuration in test files

- Each threads get their own function pointers along with stream ptr's
to be independent
- move `RuntimeTestConfig` to `RuntimeThreadConfig` in `test_types.h`

* Use datatype based calc for offsets on ptr addr

* correction for mov addr for aarch64

* Secure replacements when line starts with char '.' in asm

* Changes

- Add `TestConfigThread` - `sizes`, `offsets` params to `TestConfigStream`
- Adapted the same in kernels
    - `sizes`, `offsets` are applicable only to streams config from kernels
    - `Threads` setting in config is removed

* Calculate threads offsets and sizes per Streams in WG's

* Fix

- Calculate the sizes, offsets and its stream pointer addresses per
thread baisis
- Generate ASM Codelines per Thread per Workgroup

- Aarch64 use 'ldr' instruction than 'mov' instr
- `size_t` uses `%zu` for formatting
- Safeguard data type conversion in allocation

* Fix tests as per thread settings

* Use sysconfig for reading `CL_SIZE`

* Fix memory leaks

- `open_function` and `close_function` will be based on per `RuntimeThreadConfig` basis
- `RuntimeTestConfig` will be a ptr in `RuntimeThreadConfig`

* Avoid warnings

* Fix `CL_SIZE`

* Functionalities

- Add `bitmask` operations for
    - `roundup, rounddown, roundnearest` in general and powers of 2
    - `checking multiples and powers of 2 for bit operations`

- Add `WARN` function

- Sanitize each threads `sizes and offsets`, a `rounddown` operation based on `BYTES_PER_ITER` - read from test config file

* More detailed info for sizes and offsets in verbose mode

* Add debug info for replace lines

* Fix

- use `uin64_t` instead of `size_t`/`int64_t` for array sizes and
handling

- Fix formatting on errors for handling `PRI` for debugging statements
    - remove `STRINGIFY` for few debug, error and warn's

* Fix bitmask as per `uint64_t` set

* Formatting for `PRIu64`

* Remove redundant Variables

- `UNROLL_FACTOR` from `kernels/x86`
- `UNROLL_FACTOR` and `UOPS_PER_ITER` from `kernels/aarch64`

---------

+++ include/bitmap.h
@@ -2,6 +2,7 @@
 #ifndef BITMAP_H
 #define BITMAP_H
 
+#include <inttypes.h>
 #include <stdint.h>
 #include <stddef.h>
 
@@ -16,15 +17,15 @@
 
 typedef struct {
     BitmapDataType *data;
-    size_t          size;
-    size_t          alignment;
+    uint64_t          size;
+    uint64_t          alignment;
 } Bitmap;
 
-int create_bitmap(size_t size, size_t alignment, Bitmap* bitmap);
+int create_bitmap(uint64_t size, uint64_t alignment, Bitmap* bitmap);
 void destroy_bitmap(Bitmap *bitmap);
-int set_bit(Bitmap *bitmap, size_t index);
-int clear_bit(Bitmap *bitmap, size_t index);
-int is_bit_set(Bitmap *bitmap, size_t index);
+int set_bit(Bitmap *bitmap, uint64_t index);
+int clear_bit(Bitmap *bitmap, uint64_t index);
+int is_bit_set(Bitmap *bitmap, uint64_t index);
 void print_set_bits(const Bitmap *bitmap);
 
 #endif /* BITMAP_H */
